chore: convert rest exception mapper to Kotlin; moved to separate package
