+++ src/main/kotlin/net/atos/zac/app/util/exception/InputValidationFailedException.kt
@@ -2,7 +2,7 @@
 * SPDX-FileCopyrightText: 2024 Lifely
 * SPDX-License-Identifier: EUPL-1.2+
 */
-package net.atos.zac.app.util
+package net.atos.zac.app.util.exception
 
 import jakarta.ws.rs.BadRequestException
